[Structure] Improvements but I need to check memory consumption

+++ desafio-ios-fellipe-bravo/desafio-ios-fellipe-bravo/features/characters-list/entity/CharactersListResponse.swift
@@ -7,7 +7,7 @@ import Foundation
 
 struct CharactersListResponse: Codable {
 
-    let code: Int64
+    let code: Int
     let status, copyright, attributionText, attributionHTML: String
     let etag: String
     let data: DataClass
