+++ desafio-ios-fellipe-bravo/desafio-ios-fellipe-bravo/features/characters-list/entity/DataClass.swift
@@ -7,7 +7,7 @@ import Foundation
 
 struct DataClass: Codable {
 
-    let offset, limit, total, count: Int64
+    let offset, limit, total, count: Int
     let results: [CharacterInformation]
 
 }
